requirements.txt
flask
watchdog
sentence-transformers
faiss-cpu
pandas
numpy
requests
